Extract formatting to a seprate header
