+++ cfg/CFG.cc
@@ -1,6 +1,7 @@
 #include "cfg/CFG.h"
 #include "absl/strings/escaping.h"
 #include "absl/strings/str_split.h"
+#include "common/formatting.h"
 
 // helps debugging
 template class std::unique_ptr<sorbet::cfg::CFG>;
