+++ core/Symbols.cc
@@ -2,6 +2,7 @@
 #include "absl/strings/match.h"
 #include "common/JSON.h"
 #include "common/Levenstein.h"
+#include "common/formatting.h"
 #include "core/Context.h"
 #include "core/GlobalState.h"
 #include "core/Hashing.h"
