+++ main/options/options.cc
@@ -6,6 +6,7 @@
 #include "absl/strings/str_split.h"
 #include "common/FileOps.h"
 #include "common/Timer.h"
+#include "common/formatting.h"
 #include "core/Error.h"
 #include "core/errors/infer.h"
 #include "main/options/ConfigParser.h"
